Rename sequencing metrics module
