+++ tests/conftest.py
@@ -14,7 +14,7 @@ import time
 
 from cg_lims.EPPs.qc.sequencing_artifact_manager import SequencingArtifactManager
 from cg_lims.EPPs.qc.sequencing_quality_checker import SequencingQualityChecker
-from cg_lims.models.sequencing_metrics import SampleLaneSequencingMetrics
+from cg_lims.models.sample_lane_sequencing_metrics import SampleLaneSequencingMetrics
 
 from cg_lims.token_manager import TokenManager
 from cg_lims.status_db_api import StatusDBAPI
